requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.3
Flask-Bcrypt==1.0.1
Flask-CORS==4.0.0
PyJWT==2.7.0





